Changed script. Updated submodules.

+++ bootstrap.sh
@@ -1,8 +1,15 @@
 #!/bin/sh
 
 git submodule update --init --recursive
-DIRNAME=$(basename $(pwd))
-cd .. && ln -s $DIRNAME/.vimrc && cd -
+if [ -e ../.vimrc -a ! ../.vimrc -ef .vimrc ]
+then
+  echo "Backuping current .vimrc to .vimrc~."
+  mv ../.vimrc ../.vimrc~
+elif [ ! -e ../.vimrc ]
+then
+  DIRNAME=$(basename $(pwd))
+  cd .. && ln -s $DIRNAME/.vimrc && cd -
+fi
 # Command-T
 cd bundle/Command-T/ruby/command-t && ruby extconf.rb && make && cd -
 # YouCompleteMe
